# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  batch.BatchJobResponse:
    properties:
      completed:
        type: integer
      completed_at:
        type: string
      excluded:
        additionalProperties:
          type: boolean
        description: Files excluded from organization
        type: object
      failed:
        type: integer
      id:
        type: string
      progress:
        type: number
      results:
        additionalProperties:
          $ref: '#/definitions/contracts.BatchFileResult'
        type: object
      started_at:
        type: string
      status:
        type: string
      total_files:
        type: integer
    type: object
...
  batch.BatchScrapeRequest:
    properties:
      array_strategy:
        description: 'For Update mode: merge, replace'
        example: merge
        type: string
      destination:
        example: /path/to/output
        type: string
      files:
        items:
          type: string
        type: array
      force:
        example: false
        type: boolean
      preset:
        description: 'Merge strategy preset: conservative, gap-fill, aggressive (overrides
          scalar/array strategies)'
        example: conservative
        type: string
      scalar_strategy:
        description: 'For Update mode: prefer-nfo, prefer-scraper, preserve-existing,
          fill-missing-only'
        example: prefer-nfo
        type: string
      selected_scrapers:
        example:
        - r18dev
        - dmm
        items:
          type: string
        type: array
      strict:
        example: false
        type: boolean
      update:
        description: 'Update mode: only create/update metadata files without moving
          video files'
        example: false
        type: boolean
    required:
    - files
    type: object
  batch.BatchScrapeResponse:
    properties:
      job_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  batch.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
  batch.MovieResponse:
    properties:
      merge_stats:
        allOf:
        - $ref: '#/definitions/contracts.MergeStatistics'
        description: Merge statistics when NFO merging occurred
      movie:
        $ref: '#/definitions/models.Movie'
      provenance:
        additionalProperties:
          $ref: '#/definitions/contracts.DataSource'
        description: Field-level data source tracking
        type: object
    type: object
...
  batch.PosterCropRequest:
    properties:
      height:
        minimum: 1
        type: integer
      width:
        minimum: 1
        type: integer
      x:
        minimum: 0
        type: integer
      "y":
        minimum: 0
        type: integer
    type: object
  batch.PosterCropResponse:
    properties:
      cropped_poster_url:
        type: string
    type: object
  batch.UpdateMovieRequest:
    properties:
      movie:
        $ref: '#/definitions/models.Movie'
    required:
    - movie
    type: object
...
  contracts.BatchFileResult:
    properties:
      actress_sources:
        additionalProperties:
          type: string
        description: Actress-level source by scraper/NFO
        type: object
      data:
        description: Movie data
      ended_at:
        type: string
      error:
        type: string
      field_sources:
        additionalProperties:
          type: string
        description: Field-level source by scraper/NFO
        type: object
      file_path:
        type: string
      is_multi_part:
        type: boolean
      movie_id:
        type: string
      part_number:
        type: integer
      part_suffix:
        type: string
      started_at:
        type: string
      status:
        type: string
    type: object
  contracts.DataSource:
    properties:
      confidence:
        description: Confidence score (0.0-1.0)
        example: 0.9
        type: number
      last_updated:
        description: ISO 8601 timestamp
        example: "2024-01-15T10:30:00Z"
        type: string
      source:
        description: '"scraper" or "nfo"'
        example: nfo
        type: string
    type: object
...
  contracts.MergeStatistics:
    properties:
      conflicts_resolved:
        example: 5
        type: integer
      empty_fields:
        example: 2
        type: integer
      from_nfo:
        example: 3
        type: integer
      from_scraper:
        example: 10
        type: integer
      merged_arrays:
        example: 2
        type: integer
      total_fields:
        example: 15
        type: integer
    type: object
...
  models.Actress:
    properties:
      aliases:
        description: Pipe-separated
        type: string
      created_at:
        type: string
      dmm_id:
        description: Real DMM actress ID when available (unique only for values >
          0)
        type: integer
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      last_name:
        type: string
      thumb_url:
        type: string
      updated_at:
        type: string
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Movie:
    properties:
      actresses:
        description: Relationships
        items:
          $ref: '#/definitions/models.Actress'
        type: array
      content_id:
        type: string
      cover_url:
        description: Landscape/fanart image
        type: string
      created_at:
        type: string
      cropped_poster_url:
        description: URL to the cropped poster (persisted)
        type: string
      description:
        type: string
      director:
        type: string
      display_name:
        type: string
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: string
      label:
        description: Sub-label
        type: string
      maker:
        description: Studio/maker
        type: string
      original_filename:
        type: string
      original_title:
        description: Japanese/original language title
        type: string
      poster_url:
        description: Portrait/box art image
        type: string
      rating_score:
        type: number
      rating_votes:
        type: integer
      release_date:
        type: string
      release_year:
        type: integer
      runtime:
        description: in minutes
        type: integer
      screenshot_urls:
        items:
          type: string
        type: array
      series:
        description: Series name
        type: string
      should_crop_poster:
        description: Whether poster needs cropping from cover
        type: boolean
      source_name:
        description: Metadata
        type: string
      source_url:
        type: string
      title:
        type: string
      trailer_url:
        type: string
      translations:
        description: Translations
        items:
          $ref: '#/definitions/models.MovieTranslation'
        type: array
      updated_at:
        type: string
    type: object
  models.MovieTranslation:
    properties:
      created_at:
        type: string
      description:
        type: string
      director:
        type: string
      id:
        type: integer
      label:
        type: string
      language:
        description: 'ISO 639-1: en, ja, zh, etc.'
        type: string
      maker:
        type: string
      movie_id:
        type: string
      original_title:
        description: Japanese/original language title
        type: string
      series:
        type: string
      source_name:
        description: Which scraper provided this translation
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/javinizer/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/javinizer/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/batch/{id}:
    get:
      description: Retrieve the status of a batch scraping job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/batch.BatchJobResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/batch.ErrorResponse'
      summary: Get batch job status
      tags:
      - web
  /api/v1/batch/{id}/cancel:
    post:
      description: Cancel a running batch scraping job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/batch.ErrorResponse'
      summary: Cancel batch job
      tags:
      - web
  /api/v1/batch/{id}/movies/{movieId}:
    patch:
      consumes:
      - application/json
      description: Update a movie's metadata within a batch job's results
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: string
      - description: Updated movie data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/batch.UpdateMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/batch.MovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/batch.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/batch.ErrorResponse'
      summary: Update movie in batch job
      tags:
      - web
  /api/v1/batch/{id}/movies/{movieId}/exclude:
    post:
      description: Mark a movie in a batch job as excluded from file organization
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/batch.ErrorResponse'
      summary: Exclude movie from batch organization
      tags:
      - web
  /api/v1/batch/{id}/movies/{movieId}/poster-crop:
    post:
      consumes:
      - application/json
      description: Re-crop a temp poster for the review page using fixed-size crop
        coordinates
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: string
      - description: Crop coordinates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/batch.PosterCropRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/batch.PosterCropResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/batch.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/batch.ErrorResponse'
      summary: Update manual poster crop in batch job
      tags:
      - web
...
  /api/v1/batch/scrape:
    post:
      consumes:
      - application/json
      description: Scrape metadata for multiple movies in batch. Automatically discovers
        and includes all parts of multi-part files.
      parameters:
      - description: Batch scrape parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/batch.BatchScrapeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/batch.BatchScrapeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/batch.ErrorResponse'
      summary: Batch scrape movies
      tags:
      - web
...
schemes:
- http
- https
swagger: "2.0"