# internal/api/batch/lifecycle.go
package batch

import (
	"path/filepath"

	"github.com/gin-gonic/gin"
	"github.com/javinizer/javinizer-go/internal/logging"
	"github.com/javinizer/javinizer-go/internal/nfo"
)

// batchScrape godoc
// @Summary Batch scrape movies
// @Description Scrape metadata for multiple movies in batch. Automatically discovers and includes all parts of multi-part files.
// @Tags web
// @Accept json
// @Produce json
// @Param request body BatchScrapeRequest true "Batch scrape parameters"
// @Success 200 {object} BatchScrapeResponse
// @Failure 400 {object} ErrorResponse
// @Router /api/v1/batch/scrape [post]
func batchScrape(deps *ServerDependencies) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req BatchScrapeRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(400, ErrorResponse{Error: err.Error()})
			return
		}

		// Apply preset if specified (overrides individual strategy fields)
		if req.Preset != "" {
			var presetErr error
			req.ScalarStrategy, req.ArrayStrategy, presetErr = nfo.ApplyPreset(req.Preset, req.ScalarStrategy, req.ArrayStrategy)
			if presetErr != nil {
				c.JSON(400, ErrorResponse{Error: presetErr.Error()})
				return
			}
			logging.Infof("Applied preset '%s': scalar=%s, array=%s", req.Preset, req.ScalarStrategy, req.ArrayStrategy)
		}

		// Security: Validate all submitted files against directory security settings
		cfg := deps.GetConfig()
		for _, filePath := range req.Files {
			dir := filepath.Dir(filePath)
			if !isDirAllowed(dir, cfg.API.Security.AllowedDirectories, cfg.API.Security.DeniedDirectories) {
				// Security: Don't leak directory paths in error messages
				c.JSON(403, ErrorResponse{Error: "Access denied to requested directory"})
				return
			}
		}

		// Auto-discover sibling multi-part files
		allFiles := discoverSiblingParts(req.Files, deps.GetMatcher(), cfg)

		if len(allFiles) > len(req.Files) {
			logging.Infof("Auto-discovered %d sibling files for batch job (original: %d, total: %d)",
				len(allFiles)-len(req.Files), len(req.Files), len(allFiles))
		}

		// Create job with all files (original + discovered siblings)
		job := deps.JobQueue.CreateJob(allFiles)

		// Start processing in background - use getters for thread-safe access
		go processBatchJob(job, deps.GetRegistry(), deps.GetAggregator(), deps.MovieRepo, deps.GetMatcher(), req.Strict, req.Force, req.Update, req.Destination, deps.GetConfig(), req.SelectedScrapers, req.ScalarStrategy, req.ArrayStrategy, deps.DB)

		c.JSON(200, BatchScrapeResponse{
			JobID: job.ID,
		})
	}
}

// getBatchJob godoc
// @Summary Get batch job status
// @Description Retrieve the status of a batch scraping job
// @Tags web
// @Produce json
// @Param id path string true "Job ID"
// @Success 200 {object} BatchJobResponse
// @Failure 404 {object} ErrorResponse
// @Router /api/v1/batch/{id} [get]
func getBatchJob(deps *ServerDependencies) gin.HandlerFunc {
	return func(c *gin.Context) {
		jobID := c.Param("id")

		// GetJob() now returns a snapshot (result of GetStatus()), not a pointer
		// So we don't need to call GetStatus() again
		job, ok := deps.JobQueue.GetJob(jobID)
		if !ok {
			c.JSON(404, ErrorResponse{Error: "Job not found"})
			return
		}

		// Debug logging to trace the state
		logging.Debugf("[GET /batch/%s] Returning job with %d results, completed=%d, failed=%d",
			jobID, len(job.Results), job.Completed, job.Failed)

		var completedAt *string
		if job.CompletedAt != nil {
			str := job.CompletedAt.Format("2006-01-02T15:04:05Z07:00")
			completedAt = &str
		}

		// Transform results from worker.FileResult to BatchFileResult
		results := make(map[string]*BatchFileResult)
		for filePath, fileResult := range job.Results {
			var endedAt *string
			if fileResult.EndedAt != nil {
				str := fileResult.EndedAt.Format("2006-01-02T15:04:05Z07:00")
				endedAt = &str
			}

			results[filePath] = &BatchFileResult{
				FilePath:       fileResult.FilePath,
				MovieID:        fileResult.MovieID,
				Status:         string(fileResult.Status),
				Error:          fileResult.Error,
				FieldSources:   fileResult.FieldSources,
				ActressSources: fileResult.ActressSources,
				Data:           fileResult.Data,
				StartedAt:      fileResult.StartedAt.Format("2006-01-02T15:04:05Z07:00"),
				EndedAt:        endedAt,
				IsMultiPart:    fileResult.IsMultiPart,
				PartNumber:     fileResult.PartNumber,
				PartSuffix:     fileResult.PartSuffix,
			}
		}

		c.JSON(200, BatchJobResponse{
			ID:          job.ID,
			Status:      string(job.Status),
			TotalFiles:  job.TotalFiles,
			Completed:   job.Completed,
			Failed:      job.Failed,
			Excluded:    job.Excluded,
			Progress:    job.Progress,
			Results:     results,
			StartedAt:   job.StartedAt.Format("2006-01-02T15:04:05Z07:00"),
			CompletedAt: completedAt,
		})
	}
}

// cancelBatchJob godoc
// @Summary Cancel batch job
// @Description Cancel a running batch scraping job
// @Tags web
// @Produce json
// @Param id path string true "Job ID"
// @Success 200 {object} map[string]string
// @Failure 404 {object} ErrorResponse
// @Router /api/v1/batch/{id}/cancel [post]
func cancelBatchJob(deps *ServerDependencies) gin.HandlerFunc {
	return func(c *gin.Context) {
		jobID := c.Param("id")

		// Use GetJobPointer to get the real job (not a snapshot) so Cancel() works
		job, ok := deps.JobQueue.GetJobPointer(jobID)
		if !ok {
			c.JSON(404, ErrorResponse{Error: "Job not found"})
			return
		}

		job.Cancel()

		// Cleanup temp posters for cancelled job (batch job is gone, temp posters no longer needed)
		go cleanupJobTempPosters(jobID, deps.GetConfig().System.TempDir)

		c.JSON(200, gin.H{"message": "Job cancelled successfully"})
	}
}